requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
Werkzeug==2.3.3
gunicorn==21.2.0
Flask-Migrate==4.0.5
Flask-Limiter==3.3.0
Flask-WTF==1.1.1
Flask-Talisman==1.0.0
python-dotenv==1.0.0